requirements.txt
torch>=2.1
torch-geometric>=2.5
stable-baselines3>=2.3
wandb>=0.18
pyyaml>=6.0
numpy>=1.26
matplotlib>=3.8
pandas>=2.1
python-dotenv>=1.0
